moved preferences page out of main
