+++ main.py
@@ -6,24 +6,23 @@ from threading import Thread
 from PyQt5.QtCore import Qt, QSize, QPropertyAnimation, QEasingCurve, QPoint, QTimer, QRect, QObject, pyqtSignal
 from PyQt5.QtGui import QIcon, QPixmap, QColor
 from PyQt5.QtWidgets import QApplication, QLabel, QWidget, QMainWindow, QFrame, QPushButton, QGridLayout, QVBoxLayout, \
-    QGraphicsDropShadowEffect, QStackedWidget, QScrollArea, QScrollBar, \
-    QToolButton, QFileDialog, QSizeGrip, QProxyStyle, QStyle
+    QGraphicsDropShadowEffect, QStackedWidget, QFileDialog, QSizeGrip
 from pynput import keyboard
 
-from frontend.generic import Font, TextLabel, HyperlinkTextLabel, TextInputBox, CheckBox, Selector, Button, \
-    CheckBoxWithFunction
+from frontend.generic import Font, TextLabel, HyperlinkTextLabel, TextInputBox, Button, \
+    Icons
 from frontend.layouts import RowLayout
 from frontend.pages.bloodweb import BloodwebPage
+from frontend.pages.preferences import PreferencesPage
+from frontend.stylesheets import StyleSheets
 
 sys.path.append(os.path.dirname(os.path.realpath("backend/state.py")))
 
 from backend.config import Config
-from backend.data import Data, Unlockable
+from backend.data import Data
 from backend.state import State
 from backend.utils.text_util import TextUtil
-from frontend.stylesheets import StyleSheets
 
-# not so generic
 class TopBar(QFrame):
     def __init__(self, parent):
         super().__init__(parent)
@@ -180,214 +179,6 @@ class PageButton(QPushButton):
         self.setStyleSheet(StyleSheets.page_button)
 
         self.clicked.connect(on_click)
-
-
-# https://forum.qt.io/topic/15068/prevent-flat-qtoolbutton-from-moving-when-clicked/8
-class NoShiftStyle(QProxyStyle):
-    def pixelMetric(self, metric, option, widget):
-        if metric == QStyle.PM_ButtonShiftHorizontal or metric == QStyle.PM_ButtonShiftVertical:
-            ret = 0
-        else:
-            ret = QProxyStyle.pixelMetric(self, metric, option, widget)
-        return ret
-
-class CollapsibleBox(QWidget):
-    def __init__(self, parent, object_name, text):
-        super().__init__(parent)
-        self.layout = QVBoxLayout(self)
-        self.layout.setContentsMargins(0, 0, 0, 0)
-        self.layout.setSpacing(15)
-
-        self.toggleButton = QToolButton(self)
-        self.toggleButton.setObjectName(object_name)
-        self.toggleButton.setCheckable(True)
-        self.toggleButton.setChecked(False)
-        self.toggleButton.setFont(Font(12))
-        self.toggleButton.setStyleSheet(StyleSheets.collapsible_box_inactive)
-        self.toggleButton.setText(text)
-        self.toggleButton.setIcon(QIcon(Icons.right_arrow))
-        self.toggleButton.setIconSize(QSize(20, 20))
-        self.toggleButton.setToolButtonStyle(Qt.ToolButtonTextBesideIcon)
-        self.toggleButton.pressed.connect(self.on_pressed)
-        self.toggleButton.setStyle(NoShiftStyle())
-
-        self.layout.addWidget(self.toggleButton, alignment=Qt.AlignTop)
-
-    def on_pressed(self):
-        pass
-
-class FilterOptionsCollapsibleBox(CollapsibleBox):
-    def __init__(self, parent, object_name, on_click):
-        super().__init__(parent, object_name, "Filter Options (Click to Expand)")
-
-        # TODO clear filters button
-        # TODO new filter for positive / zero / negative tier, positive / zero / negative subtier, also sort by tier
-        self.filters = QScrollArea(self)
-        self.filters.setMinimumHeight(0)
-        self.filters.setMaximumHeight(0)
-        self.filters.setStyleSheet("""
-            QScrollArea {
-                background: rgba(0, 0, 0, 0);
-                border: 0px solid rgba(0, 0, 0, 0);
-            }""")
-
-        self.filtersLayout = QGridLayout(self.filters)
-        self.filtersLayout.setContentsMargins(0, 0, 0, 0)
-        self.filtersLayout.setSpacing(15)
-
-        # character
-        self.characterHeading = TextLabel(self.filters, "characterHeading", "Character")
-        self.filtersLayout.addWidget(self.characterHeading, 0, 0, 1, 2)
-
-        self.characterCheckBoxes = {}
-        for i, character in enumerate(Data.get_categories(True), 1):
-            checkbox = CheckBoxWithFunction(self.filters, f"{TextUtil.camel_case(character)}CharacterFilterCheckBox",
-                                            on_click)
-            checkbox.setFixedSize(25, 25)
-            self.characterCheckBoxes[character] = checkbox
-            self.filtersLayout.addWidget(checkbox, i, 0, 1, 1)
-
-            label = TextLabel(self.filters, f"{TextUtil.camel_case(character)}CharacterFilterLabel",
-                              TextUtil.title_case(character))
-            self.filtersLayout.addWidget(label, i, 1, 1, 1)
-
-        # rarity
-        self.rarityHeading = TextLabel(self.filters, "rarityHeading", "Rarity")
-        self.filtersLayout.addWidget(self.rarityHeading, 0, 2, 1, 2)
-
-        self.rarityCheckBoxes = {}
-        for i, rarity in enumerate(Data.get_rarities(), 1):
-            checkbox = CheckBoxWithFunction(self.filters, f"{TextUtil.camel_case(rarity)}RarityFilterCheckBox",
-                                            on_click)
-            checkbox.setFixedSize(25, 25)
-            self.rarityCheckBoxes[rarity] = checkbox
-            self.filtersLayout.addWidget(checkbox, i, 2, 1, 1)
-
-            label = TextLabel(self.filters, f"{TextUtil.camel_case(rarity)}RarityFilterLabel",
-                              TextUtil.title_case(rarity))
-            self.filtersLayout.addWidget(label, i, 3, 1, 1)
-
-        # type
-        self.typeHeading = TextLabel(self.filters, "typeHeading", "Type")
-        self.filtersLayout.addWidget(self.typeHeading, 0, 4, 1, 2)
-
-        self.typeCheckBoxes = {}
-        for i, unlockable_type in enumerate(Data.get_types(), 1):
-            checkbox = CheckBoxWithFunction(self.filters, f"{TextUtil.camel_case(unlockable_type)}TypeFilterCheckBox",
-                                            on_click)
-            checkbox.setFixedSize(25, 25)
-            self.typeCheckBoxes[unlockable_type] = checkbox
-            self.filtersLayout.addWidget(checkbox, i, 4, 1, 1)
-
-            label = TextLabel(self.filters, f"{TextUtil.camel_case(unlockable_type)}TypeFilterLabel",
-                              TextUtil.title_case(unlockable_type))
-            self.filtersLayout.addWidget(label, i, 5, 1, 1)
-
-        self.filtersLayout.setColumnStretch(999, 1)
-
-        self.layout.addWidget(self.filters, alignment=Qt.AlignTop)
-
-    def get_character_filters(self):
-        return [name for name, checkbox in self.characterCheckBoxes.items() if checkbox.isChecked()]
-
-    def get_rarity_filters(self):
-        return [name for name, checkbox in self.rarityCheckBoxes.items() if checkbox.isChecked()]
-
-    def get_type_filters(self):
-        return [name for name, checkbox in self.typeCheckBoxes.items() if checkbox.isChecked()]
-
-    # TODO auto resize rather than fixed number
-    def on_pressed(self):
-        if self.toggleButton.isChecked():
-            self.toggleButton.setStyleSheet(StyleSheets.collapsible_box_active)
-            self.toggleButton.setIcon(QIcon(Icons.right_arrow))
-            self.toggleButton.setText("Filter Options (Click to Expand)")
-            self.filters.setMinimumHeight(0)
-            self.filters.setMaximumHeight(0)
-        else:
-            self.toggleButton.setStyleSheet(StyleSheets.collapsible_box_inactive)
-            self.toggleButton.setIcon(QIcon(Icons.down_arrow))
-            self.toggleButton.setText("Filter Options (Click to Collapse)")
-            self.filters.setMinimumHeight(1200)
-            self.filters.setMaximumHeight(1200)
-
-class UnlockableWidget(QWidget):
-    def __init__(self, parent, unlockable: Unlockable, tier, subtier, on_unlockable_select):
-        name = TextUtil.camel_case(unlockable.name)
-
-        super().__init__(parent)
-        self.setObjectName(f"{name}Widget")
-
-        self.layout = RowLayout(self, f"{name}Layout")
-
-        self.checkBox = CheckBox(self, f"{name}CheckBox")
-        self.checkBox.setFixedSize(25, 25)
-        self.checkBox.clicked.connect(on_unlockable_select)
-        self.layout.addWidget(self.checkBox)
-
-        self.image = QLabel(self)
-        self.image.setObjectName(f"{name}Image")
-        self.image.setFixedSize(QSize(75, 75))
-        self.image.setPixmap(QPixmap(unlockable.image_path))
-        self.image.setScaledContents(True)
-        self.image.setToolTip(f"""Character: {TextUtil.title_case(unlockable.category)}
-Rarity: {TextUtil.title_case(unlockable.rarity)}
-Type: {TextUtil.title_case(unlockable.type)}""")
-        self.layout.addWidget(self.image)
-
-        self.label = QLabel(self)
-        self.label.setObjectName(f"{name}Label")
-        self.label.setFont(Font(10))
-        self.label.setStyleSheet(StyleSheets.white_text)
-        self.label.setText(unlockable.name)
-        self.layout.addWidget(self.label)
-
-        self.layout.addSpacing(250 - self.label.fontMetrics().boundingRect(self.label.text()).width())
-
-        self.tierLabel = QLabel(self)
-        self.tierLabel.setObjectName(f"{name}TierLabel")
-        self.tierLabel.setFont(Font(10))
-        self.tierLabel.setStyleSheet(StyleSheets.white_text)
-        self.tierLabel.setText("Tier")
-        self.layout.addWidget(self.tierLabel)
-
-        self.tierInput = TextInputBox(self, f"{name}TierInput", QSize(110, 40), "Enter tier", str(tier),
-                                      style_sheet=StyleSheets.tiers_input(tier))
-        self.tierInput.textEdited.connect(self.on_tier_update)
-        self.layout.addWidget(self.tierInput)
-
-        self.subtierLabel = QLabel(self)
-        self.subtierLabel.setObjectName(f"{name}TierLabel")
-        self.subtierLabel.setFont(Font(10))
-        self.subtierLabel.setStyleSheet(StyleSheets.white_text)
-        self.subtierLabel.setText("Subtier")
-        self.layout.addWidget(self.subtierLabel)
-
-        self.subtierInput = TextInputBox(self, f"{name}SubtierInput", QSize(110, 40), "Enter subtier", str(subtier),
-                                         style_sheet=StyleSheets.tiers_input(tier))
-        self.subtierInput.textEdited.connect(self.on_subtier_update)
-        self.layout.addWidget(self.subtierInput)
-
-        self.layout.addStretch(1)
-
-        self.unlockable = unlockable
-
-    def on_tier_update(self):
-        self.tierInput.setStyleSheet(StyleSheets.tiers_input(self.tierInput.text()))
-
-    def on_subtier_update(self):
-        self.subtierInput.setStyleSheet(StyleSheets.tiers_input(self.subtierInput.text()))
-
-    def setTiers(self, tier=None, subtier=None):
-        if tier is not None:
-            self.tierInput.setText(str(tier))
-            self.tierInput.setStyleSheet(StyleSheets.tiers_input(tier))
-        if subtier is not None:
-            self.subtierInput.setText(str(subtier))
-            self.subtierInput.setStyleSheet(StyleSheets.tiers_input(subtier))
-
-    def getTiers(self):
-        return int(self.tierInput.text()), int(self.subtierInput.text())
 
 '''class PromptWindow(QMainWindow):
     def __init__(self, parent, title, object_name):
@@ -561,267 +352,6 @@ class MainWindow(QMainWindow):
                                    else LeftMenuButton.max_width)
         self.animation.setEasingCurve(QEasingCurve.InOutQuint)
         self.animation.start()
-
-    def get_edit_profile(self):
-        return self.preferencesPageProfileSelector.currentText()
-
-    def update_profiles_from_config(self):
-        while self.preferencesPageProfileSelector.count() > 0:
-            self.preferencesPageProfileSelector.removeItem(0)
-        self.preferencesPageProfileSelector.addItems(Config().profile_names() + ["blank"])
-
-        while self.bloodwebPage.profileSelector.count() > 0:
-            self.bloodwebPage.profileSelector.removeItem(0)
-        self.bloodwebPage.profileSelector.addItems(Config().profile_names() + ["blank"])
-
-    def replace_unlockable_widgets(self):
-        sort_by = self.preferencesPageSortSelector.currentText()
-
-        if self.lastSortedBy == sort_by:
-            # if there was no change in ordering, don't pass in sort_by; the resulting visible list has arbitrary order
-            for widget, is_visible in Data.filter(self.preferencesPageUnlockableWidgets,
-                                                  self.preferencesPageSearchBar.text(),
-                                                  self.preferencesPageFiltersBox.get_character_filters(),
-                                                  self.preferencesPageFiltersBox.get_rarity_filters(),
-                                                  self.preferencesPageFiltersBox.get_type_filters()):
-                widget.setVisible(is_visible)
-        else:
-            # lastSortedBy was changed on this event; we need visible to be ordered to insert widgets correctly
-            for widget in self.preferencesPageUnlockableWidgets:
-                self.preferencesPageScrollAreaContentLayout.removeWidget(widget)
-
-            for widget, is_visible in Data.filter(self.preferencesPageUnlockableWidgets,
-                                                  self.preferencesPageSearchBar.text(),
-                                                  self.preferencesPageFiltersBox.get_character_filters(),
-                                                  self.preferencesPageFiltersBox.get_rarity_filters(),
-                                                  self.preferencesPageFiltersBox.get_type_filters(),
-                                                  sort_by):
-                count = self.preferencesPageScrollAreaContentLayout.count()
-                self.preferencesPageScrollAreaContentLayout.insertWidget(count - 1, widget)
-                widget.setVisible(is_visible)
-
-        self.lastSortedBy = sort_by
-        self.on_unlockable_select()
-
-    def switch_edit_profile(self):
-        if not self.ignore_profile_signals:
-            # TODO prompt: unsaved changes (save or discard)
-            config = Config()
-            for widget in self.preferencesPageUnlockableWidgets:
-                widget.setTiers(*config.preference(widget.unlockable.unique_id, self.get_edit_profile()))
-
-    def save_profile(self):
-        profile_id = self.get_edit_profile()
-        if profile_id == "blank":
-            self.show_preferences_page_save_error("You cannot save to the blank profile. "
-                                                  "Use Save As below to create a new profile.")
-        else:
-            updated_profile = Config().get_profile_by_id(profile_id).copy()
-
-            non_integer = Data.verify_tiers(self.preferencesPageUnlockableWidgets)
-            if len(non_integer) > 0:
-                self.show_preferences_page_save_error(f"There are {len(non_integer)} unlockables with invalid "
-                                                      "inputs. Inputs must be a number from -999 to 999. Changes "
-                                                      "not saved.")
-                return
-
-            for widget in self.preferencesPageUnlockableWidgets:
-                tier, subtier = widget.getTiers()
-                if tier != 0 or subtier != 0:
-                    updated_profile[widget.unlockable.unique_id] = {"tier": tier, "subtier": subtier}
-                else:
-                    updated_profile.pop(widget.unlockable.unique_id, None)
-            Config().set_profile(updated_profile)
-
-            self.show_preferences_page_save_success(f"Changes saved to profile: {profile_id}")
-            QTimer.singleShot(10000, self.hide_preferences_page_save_as_success_text)
-
-    # def rename_profile(self):
-    #     rename_prompt = PromptWindow(self, "Rename", "renamePrompt")
-    #     rename_prompt.show()
-
-    def delete_profile(self):
-        if not self.ignore_profile_signals:
-            self.ignore_profile_signals = True
-            profile_id = self.get_edit_profile()
-            if profile_id == "blank":
-                self.show_preferences_page_save_error("You cannot delete the blank profile.")
-            else:
-                # TODO "are you sure" for deleting
-                Config().delete_profile(profile_id)
-
-                self.update_profiles_from_config()
-                self.preferencesPageProfileSelector.setCurrentIndex(0)
-                self.ignore_profile_signals = False
-                self.switch_edit_profile()
-
-                self.show_preferences_page_save_success(f"Profile deleted: {profile_id}")
-            self.ignore_profile_signals = False
-
-    def show_preferences_page_save_success(self, text):
-        self.preferencesPageSaveSuccessText.setText(text)
-        self.preferencesPageSaveSuccessText.setStyleSheet(StyleSheets.pink_text)
-        self.preferencesPageSaveSuccessText.setVisible(True)
-        QTimer.singleShot(10000, self.hide_preferences_page_save_text)
-
-    def show_preferences_page_save_error(self, text):
-        self.preferencesPageSaveSuccessText.setText(text)
-        self.preferencesPageSaveSuccessText.setStyleSheet(StyleSheets.purple_text)
-        self.preferencesPageSaveSuccessText.setVisible(True)
-        QTimer.singleShot(10000, self.hide_preferences_page_save_text)
-
-    def hide_preferences_page_save_text(self):
-        self.preferencesPageSaveSuccessText.setVisible(False)
-
-    def save_as_profile(self):
-        if not self.ignore_profile_signals:
-            self.ignore_profile_signals = True # saving as new profile; don't trigger
-
-            profile_id = self.preferencesPageSaveAsInput.text()
-            self.preferencesPageSaveAsInput.setText("")
-            if profile_id == "blank":
-                self.show_preferences_page_save_as_fail_text("You cannot save to a profile named \"blank\". "
-                                                             "Try a different name.")
-            else:
-                # TODO "are you sure" for overwriting existing profile "this will overwrite an existing profile.
-                #  are you sure you want to save?"
-                new_profile = {"id": profile_id}
-
-                non_integer = Data.verify_tiers(self.preferencesPageUnlockableWidgets)
-                if len(non_integer) > 0:
-                    self.show_preferences_page_save_as_fail_text(f"There are {len(non_integer)} unlockables with "
-                                                                 "invalid inputs. Inputs must be a number from -999 to "
-                                                                 "999. Changes not saved.")
-                    return
-
-                for widget in self.preferencesPageUnlockableWidgets:
-                    tier, subtier = widget.getTiers()
-                    if tier != 0 or subtier != 0:
-                        new_profile[widget.unlockable.unique_id] = {"tier": tier, "subtier": subtier}
-
-                already_existed = Config().add_profile(new_profile)
-
-                self.update_profiles_from_config()
-                index = self.preferencesPageProfileSelector.findText(profile_id)
-                self.preferencesPageProfileSelector.setCurrentIndex(index)
-
-                if already_existed:
-                    self.show_preferences_page_save_as_success_text("Existing profile overridden with changes: "
-                                                                    f"{profile_id}")
-                else:
-                    self.show_preferences_page_save_as_success_text(f"Changes saved to new profile: {profile_id}")
-
-            self.ignore_profile_signals = False
-
-    def show_preferences_page_save_as_success_text(self, text):
-        self.preferencesPageSaveAsSuccessText.setText(text)
-        self.preferencesPageSaveAsSuccessText.setStyleSheet(StyleSheets.pink_text)
-        self.preferencesPageSaveAsSuccessText.setVisible(True)
-        QTimer.singleShot(10000, self.hide_preferences_page_save_as_success_text)
-
-    def show_preferences_page_save_as_fail_text(self, text):
-        self.preferencesPageSaveAsSuccessText.setText(text)
-        self.preferencesPageSaveAsSuccessText.setStyleSheet(StyleSheets.purple_text)
-        self.preferencesPageSaveAsSuccessText.setVisible(True)
-        QTimer.singleShot(10000, self.hide_preferences_page_save_as_success_text)
-
-    def hide_preferences_page_save_as_success_text(self):
-        self.preferencesPageSaveAsSuccessText.setVisible(False)
-
-    def selected_widgets(self):
-        return [widget for widget in self.preferencesPageUnlockableWidgets if widget.checkBox.isChecked()]
-
-    def on_unlockable_select(self):
-        num_selected = len(self.selected_widgets())
-        self.preferencesPageSelectedLabel.setText(f"{num_selected} selected")
-
-        self.preferencesPageAllCheckbox.setChecked(num_selected > 0)
-        self.preferencesPageAllCheckbox.setStyleSheet(StyleSheets.check_box_all
-                                                      if all([widget.checkBox.isChecked() for widget in
-                                                              [widget for widget in
-                                                               self.preferencesPageUnlockableWidgets
-                                                               if widget.isVisible()]])
-                                                      else StyleSheets.check_box_some)
-
-    def on_unlockable_select_all(self):
-        checked = self.preferencesPageAllCheckbox.isChecked() # whether checkbox is select all (T) or deselect all (F)
-
-        if checked:
-            for widget in self.preferencesPageUnlockableWidgets:
-                if widget.isVisible():
-                    widget.checkBox.setChecked(True)
-        else:
-            for widget in self.preferencesPageUnlockableWidgets:
-                widget.checkBox.setChecked(False)
-
-        self.on_unlockable_select()
-
-    def expand_edit(self):
-        new_pos = self.preferencesPageEditDropdownButton.mapTo(self.preferencesPage,
-                                                               self.preferencesPageEditDropdownButton.pos())
-        new_pos.setY(new_pos.y() - 250)
-        # TODO move this new_pos into an event handler
-        #  (https://stackoverflow.com/questions/15238973/how-do-i-set-a-relative-position-for-a-qt-widget)
-        #  - will work for resizing
-
-        if self.preferencesPageEditDropdownButton.isChecked():
-            self.preferencesPageEditDropdownButton.setStyleSheet(StyleSheets.collapsible_box_active)
-            self.preferencesPageEditDropdownButton.setIcon(QIcon(Icons.right_arrow))
-            self.preferencesPageEditDropdownContent.setMinimumHeight(0)
-            self.preferencesPageEditDropdownContent.setMaximumHeight(0)
-            self.preferencesPageEditDropdownContent.move(new_pos)
-        else:
-            self.preferencesPageEditDropdownButton.setStyleSheet(StyleSheets.collapsible_box_inactive)
-            self.preferencesPageEditDropdownButton.setIcon(QIcon(Icons.up_arrow))
-            self.preferencesPageEditDropdownContent.setMinimumHeight(200)
-            self.preferencesPageEditDropdownContent.setMaximumHeight(200)
-            self.preferencesPageEditDropdownContent.move(new_pos)
-
-    def on_edit_dropdown_tier(self):
-        if self.preferencesPageEditDropdownContentTierCheckBox.isChecked():
-            self.preferencesPageEditDropdownContentTierInput.setReadOnly(False)
-            text = self.preferencesPageEditDropdownContentTierInput.text()
-            self.preferencesPageEditDropdownContentTierInput.setStyleSheet(StyleSheets.tiers_input(text))
-        else:
-            self.preferencesPageEditDropdownContentTierInput.setReadOnly(True)
-            self.preferencesPageEditDropdownContentTierInput.setStyleSheet(StyleSheets.text_box_read_only)
-
-    def on_edit_dropdown_subtier(self):
-        if self.preferencesPageEditDropdownContentSubtierCheckBox.isChecked():
-            self.preferencesPageEditDropdownContentSubtierInput.setReadOnly(False)
-            text = self.preferencesPageEditDropdownContentSubtierInput.text()
-            self.preferencesPageEditDropdownContentSubtierInput.setStyleSheet(StyleSheets.tiers_input(text))
-        else:
-            self.preferencesPageEditDropdownContentSubtierInput.setReadOnly(True)
-            self.preferencesPageEditDropdownContentSubtierInput.setStyleSheet(StyleSheets.text_box_read_only)
-
-    def on_edit_dropdown_tier_input(self):
-        text = self.preferencesPageEditDropdownContentTierInput.text()
-        self.preferencesPageEditDropdownContentTierInput.setStyleSheet(StyleSheets.tiers_input(text))
-
-    def on_edit_dropdown_subtier_input(self):
-        text = self.preferencesPageEditDropdownContentSubtierInput.text()
-        self.preferencesPageEditDropdownContentSubtierInput.setStyleSheet(StyleSheets.tiers_input(text))
-
-    def on_edit_dropdown_apply(self):
-        tier = self.preferencesPageEditDropdownContentTierInput.text()
-        subtier = self.preferencesPageEditDropdownContentSubtierInput.text()
-        for widget in self.selected_widgets():
-            if self.preferencesPageEditDropdownContentTierCheckBox.isChecked():
-                widget.setTiers(tier=tier)
-            if self.preferencesPageEditDropdownContentSubtierCheckBox.isChecked():
-                widget.setTiers(subtier=subtier)
-        self.on_edit_dropdown_minimise()
-
-    def on_edit_dropdown_minimise(self):
-        if self.preferencesPageEditDropdownContentTierCheckBox.isChecked():
-            self.preferencesPageEditDropdownContentTierCheckBox.animateClick()
-        if self.preferencesPageEditDropdownContentSubtierCheckBox.isChecked():
-            self.preferencesPageEditDropdownContentSubtierCheckBox.animateClick()
-        self.preferencesPageEditDropdownContentTierInput.setText("0")
-        self.preferencesPageEditDropdownContentSubtierInput.setText("0")
-        self.preferencesPageEditDropdownButton.animateClick()
-        self.expand_edit()
 
     # bloodweb
     def get_runtime_profile(self):
@@ -1210,241 +740,14 @@ class MainWindow(QMainWindow):
         self.homePageRow4 = HomeRow(self.homePage, 4, QIcon(Icons.help), self.helpButton.on_click,
                                     "Instructions & contact details here.")
 
-        # stack: preferencesPage
-        self.preferencesPage = QWidget()
-        self.preferencesPage.setObjectName("preferencesPage")
-        self.preferencesButton.setPage(self.preferencesPage)
-
-        self.preferencesPageLayout = QGridLayout(self.preferencesPage)
-        self.preferencesPageLayout.setObjectName("preferencesPageLayout")
-        self.preferencesPageLayout.setContentsMargins(25, 25, 25, 0)
-        self.preferencesPageLayout.setSpacing(0)
-
-        self.preferencesPageScrollBar = QScrollBar(self.preferencesPage)
-        self.preferencesPageScrollBar.setObjectName("preferencesPageScrollBar")
-        self.preferencesPageScrollBar.setOrientation(Qt.Vertical)
-        self.preferencesPageScrollBar.setStyleSheet(StyleSheets.scroll_bar)
-
-        self.preferencesPageScrollArea = QScrollArea(self.preferencesPage)
-        self.preferencesPageScrollArea.setObjectName("preferencesPageScrollArea")
-        self.preferencesPageScrollArea.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOn)
-        self.preferencesPageScrollArea.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.preferencesPageScrollArea.setVerticalScrollBar(self.preferencesPageScrollBar)
-        self.preferencesPageScrollArea.setWidgetResizable(True)
-        self.preferencesPageScrollArea.setStyleSheet("""
-            QScrollArea#preferencesPageScrollArea {
-                background: transparent;
-                border: 0px;
-            }""")
-
-        self.preferencesPageScrollAreaContent = QWidget(self.preferencesPageScrollArea)
-        self.preferencesPageScrollAreaContent.setObjectName("preferencesPageScrollAreaContent")
-        self.preferencesPageScrollAreaContent.setStyleSheet("""
-            QWidget#preferencesPageScrollAreaContent {
-                background: transparent;
-            }""")
-        self.preferencesPageScrollArea.setWidget(self.preferencesPageScrollAreaContent)
-
-        self.preferencesPageScrollAreaContentLayout = QVBoxLayout(self.preferencesPageScrollAreaContent)
-        self.preferencesPageScrollAreaContentLayout.setObjectName("preferencesPageScrollAreaContentLayout")
-        self.preferencesPageScrollAreaContentLayout.setContentsMargins(0, 0, 0, 0)
-        self.preferencesPageScrollAreaContentLayout.setSpacing(15)
-
-        # save, rename, delete
-        self.preferencesPageProfileSaveRow = QWidget(self.preferencesPageScrollAreaContent)
-        self.preferencesPageProfileSaveRow.setObjectName("preferencesPageProfileSaveRow")
-        self.preferencesPageProfileSaveRowLayout = RowLayout(self.preferencesPageProfileSaveRow,
-                                                             "preferencesPageProfileSaveRowLayout")
-
-        self.preferencesPageProfileLabel = TextLabel(self.preferencesPageScrollAreaContent,
-                                                     "preferencesPageProfileLabel", "Profile", Font(12))
-
-        self.preferencesPageProfileSelector = Selector(self.preferencesPageProfileSaveRow,
-                                                       "preferencesPageProfileSelector",
-                                                       QSize(200, 40), Config().profile_names() + ["blank"])
-        self.preferencesPageProfileSelector.currentIndexChanged.connect(self.switch_edit_profile)
-
-        self.preferencesPageSaveButton = Button(self.preferencesPageProfileSaveRow, "preferencesPageSaveButton",
-                                                "Save", QSize(60, 35))
-        self.preferencesPageSaveButton.clicked.connect(self.save_profile)
-
-        # self.preferencesPageRenameButton = Button(self.preferencesPageProfileSaveRow,
-        #                                           "preferencesPageRenameButton", "Rename", QSize(75, 35))
-        # self.preferencesPageRenameButton.clicked.connect(self.rename_profile)
-
-        self.preferencesPageDeleteButton = Button(self.preferencesPageProfileSaveRow, "preferencesPageDeleteButton",
-                                                  "Delete", QSize(75, 35))
-        self.preferencesPageDeleteButton.clicked.connect(self.delete_profile)
-
-        self.preferencesPageSaveSuccessText = TextLabel(self.preferencesPageProfileSaveRow,
-                                                        "preferencesPageSaveSuccessText", "", Font(10))
-        self.preferencesPageSaveSuccessText.setVisible(False)
-
-        # save as
-        self.preferencesPageProfileSaveAsRow = QWidget(self.preferencesPageScrollAreaContent)
-        self.preferencesPageProfileSaveAsRow.setObjectName("preferencesPageProfileSaveAsRow")
-        self.preferencesPageProfileSaveAsRowLayout = RowLayout(self.preferencesPageProfileSaveAsRow,
-                                                               "preferencesPageProfileSaveAsRowLayout")
-
-        self.preferencesPageSaveAsInput = TextInputBox(self.preferencesPageProfileSaveAsRow,
-                                                       "preferencesPageSaveAsInput",
-                                                       QSize(200, 40), "Enter profile name")
-
-        self.preferencesPageSaveAsButton = Button(self.preferencesPageProfileSaveAsRow, "preferencesPageSaveAsButton",
-                                                  "Save As", QSize(80, 35))
-        self.preferencesPageSaveAsButton.clicked.connect(self.save_as_profile)
-
-        self.preferencesPageSaveAsSuccessText = TextLabel(self.preferencesPageProfileSaveAsRow,
-                                                          "preferencesPageSaveAsSuccessText",
-                                                          "", Font(10))
-        self.preferencesPageSaveAsSuccessText.setVisible(False)
-
-        # filters
-        self.preferencesPageFiltersBox = FilterOptionsCollapsibleBox(self.preferencesPageScrollAreaContent,
-                                                                     "preferencesPageFiltersBox",
-                                                                     self.replace_unlockable_widgets)
-
-        # search bar & sort
-        self.preferencesPageSearchSortRow = QWidget(self.preferencesPageScrollAreaContent)
-        self.preferencesPageSearchSortRow.setObjectName("preferencesPageSearchSortRow")
-        self.preferencesPageSearchSortRowLayout = RowLayout(self.preferencesPageSearchSortRow,
-                                                            "preferencesPageSearchSortRowLayout")
-
-        self.preferencesPageSearchBar = TextInputBox(self.preferencesPageSearchSortRow, "preferencesPageSearchBar",
-                                                     QSize(200, 40), "Search by name")
-        self.preferencesPageSearchBar.textEdited.connect(self.replace_unlockable_widgets)
-        self.preferencesPageSortLabel = TextLabel(self.preferencesPageSearchSortRow, "preferencesPageSortLabel",
-                                                  "Sort by")
-        self.preferencesPageSortSelector = Selector(self.preferencesPageSearchSortRow, "preferencesPageSortSelector",
-                                                    QSize(150, 40), Data.get_sorts())
-        self.preferencesPageSortSelector.currentIndexChanged.connect(self.replace_unlockable_widgets)
-        self.lastSortedBy = "name" # cache of last sort
-
-        # all unlockables
-        self.preferencesPageUnlockableWidgets = []
-        config = Config()
-        for unlockable in Data.get_unlockables():
-            if unlockable.category in ["unused", "retired"]:
-                continue
-            self.preferencesPageUnlockableWidgets.append(UnlockableWidget(self.preferencesPageScrollAreaContent,
-                                                                          unlockable,
-                                                                          *config.preference(unlockable.unique_id,
-                                                                                             self.get_edit_profile()),
-                                                                          self.on_unlockable_select))
-
-        # select all bar
-        self.preferencesPagePersistentBar = QWidget(self.preferencesPage)
-        self.preferencesPagePersistentBar.setObjectName("preferencesPagePersistentBar")
-        self.preferencesPagePersistentBar.setStyleSheet(f"""
-        QWidget#preferencesPagePersistentBar {{
-            border-top: 6px solid {StyleSheets.selection};
-            border-radius: 3px;           
-        }}""")
-        self.preferencesPagePersistentBar.setMinimumHeight(80)
-        self.preferencesPagePersistentBarLayout = RowLayout(self.preferencesPagePersistentBar,
-                                                            "preferencesPagePersistentBarLayout")
-
-        self.preferencesPageAllCheckbox = CheckBoxWithFunction(self.preferencesPagePersistentBar,
-                                                               "preferencesPageAllCheckbox",
-                                                               self.on_unlockable_select_all,
-                                                               style_sheet=StyleSheets.check_box_all)
-        # TODO hover over says select all / deselect all ^
-        self.preferencesPageSelectedLabel = TextLabel(self.preferencesPagePersistentBar,
-                                                      "preferencesPageSelectedLabel", "0 selected")
-        # TODO to the right, confirmation text like with saving
-
-        # edit dropdown
-        self.preferencesPageEditDropdown = QWidget(self.preferencesPagePersistentBar)
-        self.preferencesPageEditDropdown.setObjectName("preferencesPageEditDropdown")
-
-        self.preferencesPageEditDropdownLayout = QGridLayout(self.preferencesPageEditDropdown)
-        self.preferencesPageEditDropdownLayout.setContentsMargins(0, 0, 0, 0)
-        self.preferencesPageEditDropdownLayout.setSpacing(0)
-
-        self.preferencesPageEditDropdownButton = QToolButton(self.preferencesPageEditDropdown)
-        self.preferencesPageEditDropdownButton.setObjectName("preferencesPageEditDropdownButton")
-        self.preferencesPageEditDropdownButton.setCheckable(True)
-        self.preferencesPageEditDropdownButton.setChecked(False)
-        self.preferencesPageEditDropdownButton.setFont(Font(10))
-        self.preferencesPageEditDropdownButton.setStyleSheet(StyleSheets.collapsible_box_inactive)
-        self.preferencesPageEditDropdownButton.setText("Edit")
-        self.preferencesPageEditDropdownButton.setIcon(QIcon(Icons.right_arrow))
-        self.preferencesPageEditDropdownButton.setIconSize(QSize(20, 20))
-        self.preferencesPageEditDropdownButton.setToolButtonStyle(Qt.ToolButtonTextBesideIcon)
-        self.preferencesPageEditDropdownButton.setStyle(NoShiftStyle())
-        self.preferencesPageEditDropdownButton.pressed.connect(self.expand_edit)
-
-        self.preferencesPageEditDropdownContent = QWidget(self.preferencesPage)
-        self.preferencesPageEditDropdownContent.setObjectName("preferencesPageEditDropdownContent")
-        self.preferencesPageEditDropdownContent.setMinimumHeight(0)
-        self.preferencesPageEditDropdownContent.setMaximumHeight(0)
-        self.preferencesPageEditDropdownContent.setFixedWidth(260)
-        self.preferencesPageEditDropdownContent.setStyleSheet(f"""
-            QWidget#preferencesPageEditDropdownContent {{
-                background-color: {StyleSheets.passive};
-                border: 3px solid {StyleSheets.selection};
-                border-radius: 3px;
-            }}""")
-
-        self.preferencesPageEditDropdownContentLayout = QVBoxLayout(self.preferencesPageEditDropdownContent)
-        self.preferencesPageEditDropdownContentLayout.setContentsMargins(25, 25, 25, 25)
-        self.preferencesPageEditDropdownContentLayout.setSpacing(15)
-
-        self.preferencesPageEditDropdownContentTierRow = QWidget(self.preferencesPageEditDropdownContent)
-        self.preferencesPageEditDropdownContentTierRow.setObjectName("preferencesPageEditDropdownContentTierRow")
-        self.preferencesPageEditDropdownContentTierRowLayout = RowLayout(self.preferencesPageEditDropdownContentTierRow,
-                                                                         "preferencesPageEditDropdownContentTierRowLayout")
-
-        self.preferencesPageEditDropdownContentTierCheckBox = CheckBoxWithFunction(self.preferencesPageEditDropdownContentTierRow,
-                                                                                   "preferencesPageEditDropdownContentTierCheckBox",
-                                                                                   self.on_edit_dropdown_tier)
-        self.preferencesPageEditDropdownContentTierLabel = TextLabel(self.preferencesPageEditDropdownContentTierRow,
-                                                                     "preferencesPageEditDropdownContentTierLabel",
-                                                                     "Tier")
-        self.preferencesPageEditDropdownContentTierInput = TextInputBox(self.preferencesPageEditDropdownContentTierRow,
-                                                                        "preferencesPageEditDropdownContentTierInput",
-                                                                        QSize(110, 40), "Enter tier", "0",
-                                                                        style_sheet=StyleSheets.text_box_read_only)
-        self.preferencesPageEditDropdownContentTierInput.textEdited.connect(self.on_edit_dropdown_tier_input)
-        self.preferencesPageEditDropdownContentTierInput.setReadOnly(True)
-
-        self.preferencesPageEditDropdownContentSubtierRow = QWidget(self.preferencesPageEditDropdownContent)
-        self.preferencesPageEditDropdownContentSubtierRow.setObjectName("preferencesPageEditDropdownContentSubtierRow")
-        self.preferencesPageEditDropdownContentSubtierRowLayout = RowLayout(self.preferencesPageEditDropdownContentSubtierRow,
-                                                                            "preferencesPageEditDropdownContentSubtierRowLayout")
-
-        self.preferencesPageEditDropdownContentSubtierCheckBox = CheckBoxWithFunction(self.preferencesPageEditDropdownContentSubtierRow,
-                                                                                      "preferencesPageEditDropdownContentSubtierCheckBox",
-                                                                                      self.on_edit_dropdown_subtier)
-        self.preferencesPageEditDropdownContentSubtierLabel = TextLabel(self.preferencesPageEditDropdownContentSubtierRow,
-                                                                        "preferencesPageEditDropdownContentSubtierLabel",
-                                                                        "Subtier")
-        self.preferencesPageEditDropdownContentSubtierInput = TextInputBox(self.preferencesPageEditDropdownContentSubtierRow,
-                                                                           "preferencesPageEditDropdownContentSubtierInput",
-                                                                           QSize(110, 40), "Enter subtier", "0",
-                                                                           style_sheet=StyleSheets.text_box_read_only)
-        self.preferencesPageEditDropdownContentSubtierInput.textEdited.connect(self.on_edit_dropdown_subtier_input)
-        self.preferencesPageEditDropdownContentSubtierInput.setReadOnly(True)
-
-        self.preferencesPageEditDropdownContentApplyRow = QWidget(self.preferencesPageEditDropdownContent)
-        self.preferencesPageEditDropdownContentApplyRow.setObjectName("preferencesPageEditDropdownContentApplyRow")
-        self.preferencesPageEditDropdownContentApplyRowLayout = RowLayout(self.preferencesPageEditDropdownContentApplyRow,
-                                                                          "preferencesPageEditDropdownContentApplyRowLayout")
-
-        self.preferencesPageEditDropdownContentApplyButton = Button(self.preferencesPageEditDropdownContentApplyRow,
-                                                                    "preferencesPageEditDropdownContentApplyButton",
-                                                                    "Apply", QSize(70, 35))
-        self.preferencesPageEditDropdownContentApplyButton.clicked.connect(self.on_edit_dropdown_apply)
-
-        self.preferencesPageEditDropdownContentCancelButton = Button(self.preferencesPageEditDropdownContentApplyRow,
-                                                                     "preferencesPageEditDropdownContentCancelButton",
-                                                                     "Cancel", QSize(70, 35))
-        self.preferencesPageEditDropdownContentCancelButton.clicked.connect(self.on_edit_dropdown_minimise)
-
         # stack: bloodwebPage
         self.bloodwebPage = BloodwebPage()
         self.bloodwebButton.setPage(self.bloodwebPage)
         self.bloodwebPage.runButton.clicked.connect(self.run_terminate)
+
+        # stack: preferencesPage
+        self.preferencesPage = PreferencesPage(self.bloodwebPage)
+        self.preferencesButton.setPage(self.preferencesPage)
 
         # stack: helpPage
         self.helpPage = QWidget()
@@ -1737,76 +1040,6 @@ class MainWindow(QMainWindow):
         self.homePageLayout.addStretch(1)
 
         """
-        preferencesPage
-            -> scrollArea
-                -> scrollAreaContent (widget)
-                    -> labels, combobox, everything
-            -> scrollBar
-        """
-        # rows comprising the content
-        self.preferencesPageProfileSaveRowLayout.addWidget(self.preferencesPageProfileSelector)
-        self.preferencesPageProfileSaveRowLayout.addWidget(self.preferencesPageSaveButton)
-        # self.preferencesPageProfileSaveRowLayout.addWidget(self.preferencesPageRenameButton)
-        self.preferencesPageProfileSaveRowLayout.addWidget(self.preferencesPageDeleteButton)
-        self.preferencesPageProfileSaveRowLayout.addWidget(self.preferencesPageSaveSuccessText)
-        self.preferencesPageProfileSaveRowLayout.addStretch(1)
-
-        self.preferencesPageProfileSaveAsRowLayout.addWidget(self.preferencesPageSaveAsInput)
-        self.preferencesPageProfileSaveAsRowLayout.addWidget(self.preferencesPageSaveAsButton)
-        self.preferencesPageProfileSaveAsRowLayout.addWidget(self.preferencesPageSaveAsSuccessText)
-        self.preferencesPageProfileSaveAsRowLayout.addStretch(1)
-
-        self.preferencesPageSearchSortRowLayout.addWidget(self.preferencesPageSearchBar)
-        self.preferencesPageSearchSortRowLayout.addWidget(self.preferencesPageSortLabel)
-        self.preferencesPageSearchSortRowLayout.addWidget(self.preferencesPageSortSelector)
-        self.preferencesPageSearchSortRowLayout.addStretch(1)
-
-        # putting the rows into the content
-        self.preferencesPageScrollAreaContentLayout.addWidget(self.preferencesPageProfileLabel)
-        self.preferencesPageScrollAreaContentLayout.addWidget(self.preferencesPageProfileSaveRow)
-        self.preferencesPageScrollAreaContentLayout.addWidget(self.preferencesPageProfileSaveAsRow)
-        self.preferencesPageScrollAreaContentLayout.addSpacing(15)
-        self.preferencesPageScrollAreaContentLayout.addWidget(self.preferencesPageFiltersBox)
-        self.preferencesPageScrollAreaContentLayout.addWidget(self.preferencesPageSearchSortRow)
-        self.preferencesPageScrollAreaContentLayout.addSpacing(15)
-        for unlockableWidget in self.preferencesPageUnlockableWidgets:
-            self.preferencesPageScrollAreaContentLayout.addWidget(unlockableWidget)
-        self.preferencesPageScrollAreaContentLayout.addStretch(1)
-
-        # bottom persistent bar
-        self.preferencesPageEditDropdownContentTierRowLayout.addWidget(self.preferencesPageEditDropdownContentTierCheckBox)
-        self.preferencesPageEditDropdownContentTierRowLayout.addWidget(self.preferencesPageEditDropdownContentTierLabel)
-        self.preferencesPageEditDropdownContentTierRowLayout.addStretch(1)
-        self.preferencesPageEditDropdownContentTierRowLayout.addWidget(self.preferencesPageEditDropdownContentTierInput)
-
-        self.preferencesPageEditDropdownContentSubtierRowLayout.addWidget(self.preferencesPageEditDropdownContentSubtierCheckBox)
-        self.preferencesPageEditDropdownContentSubtierRowLayout.addWidget(self.preferencesPageEditDropdownContentSubtierLabel)
-        self.preferencesPageEditDropdownContentSubtierRowLayout.addStretch(1)
-        self.preferencesPageEditDropdownContentSubtierRowLayout.addWidget(self.preferencesPageEditDropdownContentSubtierInput)
-
-        self.preferencesPageEditDropdownContentApplyRowLayout.addStretch(1)
-        self.preferencesPageEditDropdownContentApplyRowLayout.addWidget(self.preferencesPageEditDropdownContentApplyButton)
-        self.preferencesPageEditDropdownContentApplyRowLayout.addWidget(self.preferencesPageEditDropdownContentCancelButton)
-
-        self.preferencesPageEditDropdownContentLayout.addWidget(self.preferencesPageEditDropdownContentTierRow)
-        self.preferencesPageEditDropdownContentLayout.addWidget(self.preferencesPageEditDropdownContentSubtierRow)
-        self.preferencesPageEditDropdownContentLayout.addWidget(self.preferencesPageEditDropdownContentApplyRow)
-
-        self.preferencesPageEditDropdownLayout.addWidget(self.preferencesPageEditDropdownButton, 0, 0, 1, 1)
-
-        self.preferencesPagePersistentBarLayout.addWidget(self.preferencesPageAllCheckbox)
-        self.preferencesPagePersistentBarLayout.addWidget(self.preferencesPageSelectedLabel)
-        self.preferencesPagePersistentBarLayout.addWidget(self.preferencesPageEditDropdown)
-        self.preferencesPagePersistentBarLayout.addStretch(1)
-
-        # assembling it all together
-        self.preferencesPageLayout.addWidget(self.preferencesPageScrollArea, 0, 0, 1, 1)
-        self.preferencesPageLayout.addWidget(self.preferencesPageScrollBar, 0, 1, 1, 1)
-        self.preferencesPageLayout.addWidget(self.preferencesPagePersistentBar, 1, 0, 1, 1, Qt.AlignBottom)
-        self.preferencesPageLayout.setRowStretch(0, 1)
-        self.preferencesPageLayout.setColumnStretch(0, 1)
-
-        """
         helpPage
         """
         self.helpPageContactDiscordRowLayout.addWidget(self.helpPageContactDiscordIcon)
@@ -1861,34 +1094,12 @@ class MainWindow(QMainWindow):
         self.settingsPageLayout.addWidget(self.settingsPageSaveRow)
         self.settingsPageLayout.addStretch(1)
 
-        self.preferencesPageSortSelector.setCurrentIndex(1) # self.lastSortedBy becomes "character"
-
         self.emitter.prestige.connect(self.bloodwebPage.on_prestige_signal)
         self.emitter.bloodpoint.connect(self.bloodwebPage.on_bloodpoint_signal)
         self.emitter.terminate.connect(self.state.terminate)
         self.emitter.toggle_text.connect(self.toggle_run_terminate_text)
 
         self.show()
-
-class Icons:
-    __base = "assets/images/icons"
-    icon = "assets/images/inspo1.png"
-    splash = "assets/images/splash.png"
-    minimize = __base + "/icon_minimize.png"
-    restore = __base + "/icon_restore.png"
-    maximize = __base + "/icon_maximize.png"
-    close = __base + "/icon_close.png"
-    menu = __base + "/icon_menu.png"
-    home = __base + "/icon_home.png"
-    preferences = __base + "/icon_preferences.png"
-    settings = __base + "/icon_settings.png"
-    bloodweb = __base + "/icon_graph.png"
-    help = __base + "/icon_help.png"
-    down_arrow = __base + "/icon_down_arrow.png"
-    right_arrow = __base + "/icon_right_arrow.png"
-    up_arrow = __base + "/icon_up_arrow.png"
-    discord = __base + "/icon_discord.png"
-    twitter = __base + "/icon_twitter.png"
 
 # https://stackoverflow.com/questions/26746379/how-to-signal-slots-in-a-gui-from-a-different-process
 # https://stackoverflow.com/questions/34525750/mainwindow-object-has-no-attribute-connect
